Add files via upload
This is a Mini project in python. Spell-checker is about checking the spelling of the city names and correcting it. I used a fuzzy-wuzzy library which is helpful in finding the similarities between words, say Hyderabad and hyderbad will have a similarity of 95%. I have created a frequency list of 40,000 city names, whenever the city name was entered it was checked for the spelling and it go for correction. The word in frequency list which has more similarity with the word entered will be replaced in the sentence. I also setup the model in a way that it rejects the abbreviations and the stop words.